postal_code_readerをモジュール化
